+++ CHANGELOG.md
@@ -72,6 +72,8 @@
 ### Removed
 ### Fixed
 - Playlist Tools\Pools: 'SORTED BY' expressions not working properly on 'Custom pool...' entry.
+- Export settings: added missing buttons file on zip.
+- Import settings: multiple fixes to extra data files importing.
 
 ## [5.1.0] - 2025-09-29
 ### Added